Add check_mode support to route53_zone
Allow route53_zone to work in check_mode

+++ lib/ansible/modules/cloud/amazon/route53_zone.py
@@ -144,7 +144,8 @@ def main():
         vpc_id=dict(default=None),
         vpc_region=dict(default=None),
         comment=dict(default='')))
-    module = AnsibleModule(argument_spec=argument_spec)
+    module = AnsibleModule(argument_spec=argument_spec,
+                           supports_check_mode=True)
 
     if not HAS_BOTO:
         module.fail_json(msg='boto required for this module')
@@ -219,15 +220,17 @@ def main():
         module.exit_json(changed=False, set=record)
 
     elif state == 'present':
-        result = conn.create_hosted_zone(zone_in, **record)
-        hosted_zone = result['CreateHostedZoneResponse']['HostedZone']
-        zone_id = hosted_zone['Id'].replace('/hostedzone/', '')
-        record['zone_id'] = zone_id
+        if not module.check_mode:
+            result = conn.create_hosted_zone(zone_in, **record)
+            hosted_zone = result['CreateHostedZoneResponse']['HostedZone']
+            zone_id = hosted_zone['Id'].replace('/hostedzone/', '')
+            record['zone_id'] = zone_id
         record['name'] = zone_in
         module.exit_json(changed=True, set=record)
 
     elif state == 'absent' and zone_in in zones:
-        conn.delete_hosted_zone(zones[zone_in])
+        if not module.check_mode:
+            conn.delete_hosted_zone(zones[zone_in])
         module.exit_json(changed=True)
 
     elif state == 'absent':
